Extract a Scopable concern
So future custom participatory spaces don't need to duplicate the code
if they want to use scopes.
