+++ decidim-core/app/models/decidim/participatory_process.rb
@@ -9,6 +9,7 @@ module Decidim
   class ParticipatoryProcess < ApplicationRecord
     include Decidim::HasAttachments
     include Decidim::Publicable
+    include Decidim::Scopable
 
     belongs_to :organization,
                foreign_key: "decidim_organization_id",
@@ -17,10 +18,6 @@ module Decidim
                foreign_key: "decidim_participatory_process_group_id",
                class_name: "Decidim::ParticipatoryProcessGroup",
                inverse_of: :participatory_processes,
-               optional: true
-    belongs_to :scope,
-               foreign_key: "decidim_scope_id",
-               class_name: "Decidim::Scope",
                optional: true
     has_many :steps,
              -> { order(position: :asc) },
@@ -59,14 +56,6 @@ module Decidim
 
     def hashtag
       attributes["hashtag"].to_s.delete("#")
-    end
-
-    # Gets the children scopes for the participatory process.
-    # If it is a global process, returns the organization's top scopes.
-    #
-    # Returns an ActiveRecord::Relation.
-    def subscopes
-      scope ? scope.children : organization.top_scopes
     end
   end
 end
